refactor: reorganize error handling module structure
