+++ src/error/index.ts
@@ -1,0 +1,6 @@
+/**
+ * Error handling types and classes for ts-inspect library.
+ */
+
+export { TsInspectError } from "./error-class.ts";
+export { errorTypeToMessage, type TsInspectErrorType } from "./error-type.ts";
