+++ src/main/orchestrator/tsconfig/resolver.test.ts
@@ -1,6 +1,6 @@
 import assert from "node:assert";
 import { describe, it } from "node:test";
-import { TsInspectError } from "../../../error.ts";
+import { TsInspectError } from "../../../error/index.ts";
 import { resolveProjectPath } from "./resolver.ts";
 
 describe("main/orchestrator/tsconfig/resolver", () => {
